Product added

+++ models.py
@@ -1,5 +1,5 @@
 from database import Base
-from sqlalchemy import Column, Integer, String, Boolean, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, Boolean, ForeignKey, Text, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy_utils import ChoiceType
 
@@ -41,7 +41,7 @@ class Product(Base):
     __tablename__ = 'products'
     id = Column(Integer, primary_key=True)
     name = Column(String(50), nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(Float, nullable=False)
     orders = relationship("Order", back_populates="product")
 
     def __repr__(self):
